+++ modules/servers/apnscp/hooks.php
@@ -5,7 +5,29 @@ require_once('lib/Helper.php');
 add_hook('ClientAreaPageProductDetails', 1, function ($vars) {
 
     $server = $vars['serverdata'];
-    $ip     = WHMCS\Utility\Environment\CurrentUser::getIP();
+
+    // WHMCS MAGIC!!!1
+    $ca           = new WHMCS\ClientArea();
+    $legacyClient = new WHMCS\Client($ca->getClient());
+    $service      = new WHMCS\Service($vars['serviceid'], $legacyClient->getID());
+    $product      = WHMCS\Product\Product::where('id', $service->getData('packageid'))->first();
+
+    if ($product->module !== "apnscp")
+    {
+        return null;
+    }
+
+    if (version_compare(WHMCS\Config\Setting::getValue("Version"), '8', '<'))
+    {
+        $ip = WHMCS\Utility\Environment\CurrentUser::getIP();
+    }
+    else
+    {
+        $user = (new WHMCS\Authentication\CurrentUser())->user();
+        $ip = $user->currentIp();
+    }
+
+    //    $ip     = '2.2.2.2'; // Only for debug...
 
     $knownJails = [
         'f2b-sshd'         => 'SSH (f2b-sshd)',
@@ -16,17 +38,10 @@ add_hook('ClientAreaPageProductDetails', 1, function ($vars) {
         'f2b-evasive'      => 'Apache Mod Evasive (f2b-evasive)',
         'f2b-spambots'     => 'Irregular Mail Patterns (f2b-spambots)',
         'f2b-recidive'     => 'Repeatedly getting banned (f2b-recidive)',
+        'f2b-pgsql'        => 'pgSQL (f2b-pgsql)',
+        'f2b-mysql'        => 'MySQL (f2b-mysql)',
+        'f2b-malware'      => 'Malware Scans (f2b-malware)',
     ];
-
-    // WHMCS MAGIC!!!
-    $ca           = new WHMCS\ClientArea();
-    $legacyClient = new WHMCS\Client($ca->getClient());
-    $service      = new WHMCS\Service($vars['serviceid'], $legacyClient->getID());
-
-    if ($service->product->module != "apnscp")
-    {
-        return null;
-    }
 
     $serverParams = [
         'serverhttpprefix' => $server['secure'] === 'on' ? 'https' : 'http',
@@ -40,25 +55,24 @@ add_hook('ClientAreaPageProductDetails', 1, function ($vars) {
 
     $result = apnscp_checkIP($serverParams);
 
-    if ($result)
+    if ($result['jails'])
     {
-        $jails = [];
-        if (is_array($result))
+        if (is_array($result['jails']))
         {
-            foreach ($result as $jail)
+            foreach ($result['jails'] as $jail)
             {
                 $jails[] = $knownJails[ $jail ];
             }
         }
         else
         {
-            $jails[] = $result;
+            $jails[] = $result['jails'];
         }
 
-        return ['apiscp_banned' => true, 'apiscp_jails' => $jails];
+        return ['apisVars' => ['is_banned' => true, 'jails' => $jails, 'ip' => $ip, 'rampart_enabled' => $result['rampart_enabled']]];
     }
 
-    return ['apiscp_banned' => false, 'apiscp_jails' => []];
+    return ['apisVars' => ['is_banned' => false, 'jails' => [], 'ip' => $ip, 'rampart_enabled' => false]];
 });
 
 add_hook('ClientAreaPrimarySidebar', 1, function ($sidebar) {
@@ -67,12 +81,12 @@ add_hook('ClientAreaPrimarySidebar', 1, function ($sidebar) {
         return null;
     }
     $service = Menu::context("service");
-    if ($service instanceof WHMCS\Service\Service && $service->product->module != "apnscp")
+    if ($service instanceof WHMCS\Service\Service && $service->product->module !== "apnscp")
     {
         return null;
     }
 
-    $sidebar->getChild("Service Details Actions")->addChild("Login to Panel", array("uri" => "clientarea.php?action=productdetails&id=" . $service->id . "&dosinglesignon=1", "label" => 'Login to Panel', "attributes" => array("target" => "_blank"), "disabled" => $service->status != "Active", "order" => 1));
+    $sidebar->getChild("Service Details Actions")->addChild("Login to LiPanel", array("uri" => "clientarea.php?action=productdetails&id=" . $service->id . "&dosinglesignon=1", "label" => 'Login to LiPanel', "attributes" => array("target" => "_blank"), "disabled" => $service->status != "Active", "order" => 1));
 });
 
 function apnscp_checkIP(array $params)
@@ -93,12 +107,16 @@ function apnscp_checkIP(array $params)
             $jails = $client->rampart_banned_services($clientIp);
             $client->rampart_unban($clientIp);
 
-            $session_id = $client->admin_hijack($site_domain, $site_admin, 'SOAP');
-            $newClient  = Connector::create_client($apnscp_apikey, $apnscp_apiendpoint, $session_id, $clientIp);
+            $rampartCheck    = $client->admin_collect(['rampart.enabled'], null, [$site_domain]);
+            $rampart_enabled = $rampartCheck[ $site_domain ]['rampart']['enabled'] === 1;
 
-            $newClient->rampart_whitelist($clientIp);
+            // Disabled whitelisting to avoid conflicts, will instruct customer to whitelist themselves!
+//            $session_id = $client->admin_hijack($site_domain, $site_admin, 'SOAP');
+//            $newClient  = Connector::create_client($apnscp_apikey, $apnscp_apiendpoint, $session_id, $clientIp);
+//
+//            $newClient->rampart_whitelist($clientIp);
 
-            return $jails;
+            return ['jails' => $jails, 'rampart_enabled' => $rampart_enabled];
         }
 
         return false;
@@ -119,3 +137,25 @@ function apnscp_checkIP(array $params)
 
     return false;
 }
+/**
+ * This hook will automatically purge Suspended accounts in ApisCP that were suspended instead of being terminated.
+ * Go read the readme before enabling this feature!
+ */
+
+//add_hook('DailyCronJob', 1, function () {
+//
+//    $servers = WHMCS\Product\Server::where('type', 'apnscp')->where('active', 1)->where('disabled', 0)->get();
+//    foreach ($servers as $server)
+//    {
+//        $apnscp_apiendpoint = $server->serverhttpprefix . '://' . $server->serverhostname . ':' . $server->serverport ?: 2083;
+//        $apnscp_apikey      = decrypt($server->serverpassword);
+//
+//        $adminId = \session_id();
+//        $client  = Connector::create_client($apnscp_apikey, $apnscp_apiendpoint, $adminId);
+//
+//        $opts['since'] = "30 days ago";
+//        $opts['match'] = "Customer Requested Cancellation";
+//        $client->admin_delete_site(null, $opts);
+//    }
+//
+//});